Modification de la section About: Rajout des skills / animation
